requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
boto3==1.14.62
botocore==1.17.62
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.3.0
django-heroku==0.3.1
django-widget-tweaks==1.4.8
docutils==0.15.2
isort==5.5.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
django-storages==1.10.1
